Bug 1342012 - Rename ModuleScript source files to LoadedScript r=smaug
