+++ dom/script/LoadedScript.cpp
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-#include "ModuleScript.h"
+#include "LoadedScript.h"
 
 #include "mozilla/HoldDropJSObjects.h"
