+++ dom/script/LoadedScript.h
